feat: Implement gRPC services for job and geographical data, including proto definitions, controllers, and an authentication interceptor.

+++ src/main/java/com/jobify/api/dto/CityDTO.java
@@ -1,13 +1,12 @@
 package com.jobify.api.dto;
 
 import lombok.Data;
-import java.math.BigDecimal;
 
 @Data
 public class CityDTO {
     private Long id;
     private String name;
-    private BigDecimal latitude;
-    private BigDecimal longitude;
+    private Double latitude;
+    private Double longitude;
     private Long population;
 }
